move base64 functions to helpers
